requirements_vercel.txt
# Minimal requirements for Vercel deployment
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
google-generativeai==0.8.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
PyPDF2==3.0.1
nltk==3.8.1
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2
